// frontend/src/booking_activity/interfaces/ICruiseTrip.ts
export interface CruiseTripInterface {
    ID?: number;
    CruiseTripName?: string;
    Deets?: string;
    StartDate?: Date | string;
    EndDate?: Date | string;
    RoutesID?: number;
    PlanImg?: string;
    PlanPrice?: number;
    ParticNum?: number;
    ShipID?: number;
    EmployeesID?: number;
}